chore: demo cleanup (#9)
