+++ lib/credentials.dart
@@ -1,0 +1,9 @@
+class Credentials {
+  Credentials._internal();
+
+  static const applicationID = "latency";
+  static const searchOnlyKey = "927c3fe76d4b52c5a2912973f35a3077";
+  static const hitsIndex = "STAGING_native_ecom_demo_products";
+  static const suggestionsIndex =
+      "STAGING_native_ecom_demo_products_query_suggestions";
+}
