+++ lib/data/algolia_client.dart
@@ -1,18 +1,21 @@
 import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter_ecom_demo/model/query.dart';
 import 'package:http/http.dart' as http;
-import 'package:flutter_ecom_demo/domain/query.dart';
 
 /// API client for Algolia.
 class AlgoliaAPIClient extends http.BaseClient {
+  static final http.Client defaultHttpClient = http.Client();
+
   final String appID;
   final String apiKey;
   final String indexName;
-  final http.Client _client = http.Client();
+  final http.Client _client = defaultHttpClient;
 
   AlgoliaAPIClient(this.appID, this.apiKey, this.indexName);
 
+  @override
   Future<http.StreamedResponse> send(http.BaseRequest request) {
     request.headers['content-type'] = "application/json";
     request.headers['X-Algolia-API-Key'] = apiKey;
@@ -22,17 +25,17 @@ class AlgoliaAPIClient extends http.BaseClient {
 
   /// Run a search query and get a response.
   Future<Map<dynamic, dynamic>> search(Query query) async {
-    var url = Uri.https('$appID-dsn.algolia.net', '1/indexes/$indexName/query');
+    final url = Uri.https('$appID-dsn.algolia.net', '1/indexes/$indexName/query');
     final request = http.Request("post", url);
     request.body = '{"params": "${query.toParams()}"}';
-    log('[Request]: ${request.body}');
+    _log('[Request]: ${request.body}');
     final streamedResponse = await send(request);
     final response = await http.Response.fromStream(streamedResponse);
-    log('[Response]: ${response.body}');
+    _log('[Response]: ${response.body}');
     return jsonDecode(utf8.decode(response.bodyBytes)) as Map;
   }
 
-  void log(String string) {
+  void _log(String string) {
     if (kDebugMode) print(string);
   }
 }
