--- lib/domain/highlighted_string.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HighlightedString {
-
-  String string;
-  bool isHighlighted;
-
-  HighlightedString(this.string, this.isHighlighted);
-
-  @override
-  String toString() {
-    return (isHighlighted ? ">" : "") + "$string";
-  }
-
-}
