+++ lib/main.dart
@@ -1,38 +1,26 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_ecom_demo/ui/home_screen.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter_ecom_demo/ui/theme_colors.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_ecom_demo/ui/app_theme.dart';
+import 'package:flutter_ecom_demo/ui/screens/home/home_screen.dart';
+
 import 'firebase_options.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp(
-    options: DefaultFirebaseOptions.currentPlatform,
-  );
-  runApp(SWApp());
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
+  runApp(const SWApp());
 }
 
+/// S&W Fashion App entry point.
 class SWApp extends StatelessWidget {
+  const SWApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    const primaryColor = ThemeColors.darkBlue;
     return MaterialApp(
       title: 'Spencer & Williams Fashion',
-      theme: ThemeData(
-          fontFamily: 'Inter',
-          primaryColor: primaryColor,
-          scaffoldBackgroundColor: Colors.white,
-          buttonTheme: const ButtonThemeData(textTheme: ButtonTextTheme.primary),
-          appBarTheme: const AppBarTheme(
-            backgroundColor: Colors.white,
-            foregroundColor: primaryColor,
-            //toolbarTextStyle: TextStyle(color: primaryColor)
-          ),
-          textButtonTheme: TextButtonThemeData(
-              style: TextButton.styleFrom(primary: primaryColor))),
+      theme: AppTheme.buildLightTheme(),
       home: const HomeScreen(),
     );
   }
 }
-
-
