+++ lib/model/highlighted_string.dart
@@ -1,0 +1,11 @@
+class HighlightedString {
+  HighlightedString(this.string, this.isHighlighted);
+
+  String string;
+  bool isHighlighted;
+
+  @override
+  String toString() {
+    return (isHighlighted ? ">" : "") + string;
+  }
+}
