+++ lib/model/query_suggestion.dart
@@ -1,14 +1,11 @@
-
 class QuerySuggestion {
+  QuerySuggestion(this.query, this.highlighted);
 
   String query;
   String? highlighted;
-
-  QuerySuggestion(this.query, this.highlighted);
 
   static QuerySuggestion fromJson(Map<String, dynamic> json) {
     return QuerySuggestion(
         json["query"], json["_highlightResult"]["query"]["value"]);
   }
-
 }
