+++ lib/ui/screens/home/components/home_banner_view.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+
+class HomeBannerView extends StatelessWidget {
+  const HomeBannerView({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      alignment: AlignmentDirectional.centerStart,
+      children: [
+        Container(
+          alignment: Alignment.center,
+          child: ColorFiltered(
+            colorFilter: ColorFilter.mode(
+                Colors.black.withOpacity(0.33), BlendMode.srcOver),
+            child: Image.asset(
+              'assets/images/banner.jpg',
+              height: 128,
+              width: MediaQuery.of(context).size.width,
+              fit: BoxFit.cover,
+            ),
+          ),
+        ),
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text('New\nCollection'.toUpperCase(),
+                  style: Theme.of(context).textTheme.headline4?.copyWith(
+                      fontWeight: FontWeight.bold, color: Colors.white)),
+              Text('Spring/Summer 2021'.toUpperCase(),
+                  style: Theme.of(context).textTheme.subtitle2?.copyWith(
+                      fontWeight: FontWeight.bold, color: Colors.white)),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
