--- lib/ui/theme_colors.dart
@@ -1,10 +1,0 @@
-import 'dart:ui';
-
-class ThemeColors {
-  ThemeColors._internal();
-
-  static const Color nebula = Color(0xFF5468FF);
-  static const Color darkBlue = Color(0xFF21243D);
-  static const Color darkPink = Color(0xFFAA086C);
-  static const Color vividOrange = Color(0xFFE8600A);
-}
