+++ lib/ui/widgets/product_card_view.dart
@@ -1,20 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_ecom_demo/domain/product.dart';
-import 'package:flutter_ecom_demo/ui/theme_colors.dart';
-import 'package:flutter_ecom_demo/ui/widgets/color_indicator.dart';
-import 'package:flutter_ecom_demo/ui/widgets/rating_display.dart';
+import 'package:flutter_ecom_demo/model/product.dart';
+import 'package:flutter_ecom_demo/ui/app_theme.dart';
+import 'package:flutter_ecom_demo/ui/widgets/color_indicator_view.dart';
+import 'package:flutter_ecom_demo/ui/widgets/rating_view.dart';
 
 class ProductCardView extends StatelessWidget {
   const ProductCardView(
       {Key? key,
       required this.product,
       this.imageAlignment = Alignment.center,
-      this.onProductPressed})
+      this.onTap})
       : super(key: key);
 
   final Product product;
   final Alignment imageAlignment;
-  final ValueChanged<String>? onProductPressed;
+  final Function(String)? onTap;
 
   @override
   Widget build(BuildContext context) {
@@ -24,9 +24,7 @@ class ProductCardView extends StatelessWidget {
     final crossedValue =
         (product.price?.onSales ?? false) ? product.price?.value : null;
     return GestureDetector(
-      onTap: () {
-        onProductPressed?.call(product.objectID!);
-      },
+      onTap: () => onTap?.call(product.objectID!),
       child: SizedBox(
         width: 150,
         child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
@@ -44,7 +42,7 @@ class ProductCardView extends StatelessWidget {
                   child: Text(" ON SALE ${product.price?.discountLevel}% ",
                       style: Theme.of(context).textTheme.caption?.copyWith(
                           color: Colors.white,
-                          backgroundColor: ThemeColors.darkPink)),
+                          backgroundColor: AppTheme.darkPink)),
                 )
             ],
           ),
@@ -63,7 +61,7 @@ class ProductCardView extends StatelessWidget {
                   style: Theme.of(context).textTheme.bodyText2)),
           Padding(
             padding: const EdgeInsets.symmetric(vertical: 2.0),
-            child: ColorIndicator(product: product),
+            child: ColorIndicatorView(product: product),
           ),
           Row(
             children: [
@@ -73,7 +71,7 @@ class ProductCardView extends StatelessWidget {
                   softWrap: false,
                   style: Theme.of(context).textTheme.bodyText2?.copyWith(
                       fontWeight: FontWeight.bold,
-                      color: ThemeColors.vividOrange)),
+                      color: AppTheme.vividOrange)),
               if (crossedValue != null)
                 Padding(
                   padding: const EdgeInsets.only(left: 8.0),
@@ -88,7 +86,7 @@ class ProductCardView extends StatelessWidget {
                 ),
             ],
           ),
-          RatingDisplay(
+          RatingView(
               value: product.reviews?.rating?.toInt() ?? 0,
               reviewsCount: product.reviews?.count?.toInt() ?? 0),
         ]),
